Backend cập nhật tính năng về phòng, suất chiếu

+++ src/main/java/com/cinema/service/RoomService.java
@@ -8,6 +8,7 @@ import com.cinema.mapper.RoomMapper;
 import com.cinema.mapper.SeatMapper;
 import com.cinema.repository.RoomRepository;
 import com.cinema.repository.SeatRepository;
+import com.cinema.repository.ShowtimeRepository;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
@@ -26,6 +27,7 @@ public class RoomService {
     SeatRepository seatRepository;
     RoomMapper roomMapper;
     SeatMapper seatMapper;
+    ShowtimeRepository showtimeRepository;
 
     public List<RoomDto> getAllRooms() {
         return roomRepository.findAll().stream()
@@ -82,8 +84,12 @@ public class RoomService {
         return roomMapper.toDto(room);
     }
 
+    // Xóa phòng
     @Transactional
     public void deleteRoom(Integer id) {
+        if (showtimeRepository.existsByRoomId(id)) {
+            throw new RuntimeException("Không thể xóa phòng vì phòng có suất chiếu");
+        }
         Room room = roomRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Không tìm thấy phòng với ID: " + id));
         roomRepository.delete(room);
